added cleaner implementation of test socket test cases (will have to port over for windows and mac)
